requirements.txt
# Core (all phases)
anthropic>=0.40.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
httpx>=0.27.0

# Phase 1+ (MCP server/client)
mcp>=1.0.0

# Phase 2+ (knowledge graph)
networkx>=3.2

# Phase 4+ (semantic layer)
sentence-transformers>=3.0.0
numpy>=1.24.0
